Splits DataGenerator into multiple files according to class
