+++ DataGenerator/FileLoader.py
@@ -1,0 +1,96 @@
+class FileLoader:
+    import os
+    import wget
+    import zipfile
+    from tqdm import tqdm
+    TextPath = ""
+    LetterPath = ""
+    TextDownloadURL = []
+    LetterDownloadURL = ""
+    TempDownloadLetterPath = "./letters.zip"
+    TextFileQueue = []
+    TextFileStream = []
+    LetterPaths = {}
+
+    def __init__(self, textPath, letterPath, letterDownloadURL, textDownloadURL):
+        self.TextPath = textPath
+        self.LetterPath = letterPath
+        self.LetterDownloadURL = letterDownloadURL
+        for URL in textDownloadURL:
+            self.TextDownloadURL.append(URL)
+
+    def CheckAndCreatePaths(self):
+        print("Checking and creating file paths ... ", end="")
+        if not self.os.path.isdir(self.TextPath):
+            self.ImportTexts()
+        if not self.os.path.isdir(self.LetterPath):
+            self.ImportLetters()
+
+    def ImportTexts(self):
+        print("Texts not found ...")
+        self.os.makedirs(self.TextPath)
+        for file in self.TextDownloadURL:
+            print("Downloading " + file[0])
+            self.wget.download(file[1], self.TextPath + file[0] + ".txt")
+            print()
+
+    def ImportLetters(self):
+        print("Letters not found ... ", end="")
+        file = None
+        if not self.os.path.isfile(self.TempDownloadLetterPath):
+            print("Letters not downloaded ... Beginning download of letters ... ")
+            file = self.wget.download(
+                self.LetterDownloadURL, self.TempDownloadLetterPath)
+            print()
+            print("Letters downloaded ... ", end="")
+        else:
+            file = self.TempDownloadLetterPath
+            print("Letters zip file already downloaded ... ", end="")
+
+        print("Beginning extraction ... ")
+        # self.os.makedirs(self.LetterPath)
+        with self.zipfile.ZipFile(file, "r") as zf:
+            fileList = []
+            for file in zf.namelist():
+                if 'train' not in file:
+                    fileList.append(file)
+            for file in self.tqdm(iterable=fileList, total=len(fileList)):
+                zf.extract(member=file, path=(self.LetterPath + "../"))
+
+        # Move up one folder
+        print("Done")
+
+    def LoadLetterPaths(self):
+        print("Loading letter paths ... ", end="")
+        self.TextFileQueue = self.os.listdir(self.TextPath)
+        if not self.TextFileQueue:
+            raise Exception("Text files not found!")
+        self.TextFileStream = open(
+            self.TextPath + self.TextFileQueue[0], 'r', encoding='utf-8')
+        self.TextFileQueue.pop(0)
+        print("Done")
+
+    def GatherLetterPaths(self):
+        print("Gathering letter paths")
+        self.LetterPaths = {}
+        for i in self.tqdm([*range(65, 91), *range(97, 122)]):
+            self.LetterPaths[chr(i)] = {}
+            hexLetter = hex(i).split('x')[-1]
+            self.LetterPaths[chr(i)]['hex'] = hexLetter
+
+        print("Bind letter paths to a letters")
+        # for each letter (lowercase/uppercase)
+        for n in self.tqdm(self.LetterPaths):
+            self.LetterPaths[n]['paths'] = []
+            # for each hsf num
+            for i in range(0, 8):
+                hsfPath = self.LetterPath + \
+                    self.LetterPaths[n]['hex'] + "/hsf_" + str(i) + "/"
+                if self.os.path.isdir(hsfPath):
+                    pictures = self.os.listdir(hsfPath)
+                    # for each picture in directory
+                    for t in pictures:
+                        self.LetterPaths[n]['paths'].append(hsfPath + t)
+
+    def Finish(self):
+        self.TextFileStream.close()
